Homework 7

+++ database/migrations/2017_01_21_194552_create_article_table_and_article_user_table.php
@@ -20,7 +20,7 @@ class CreateArticleTableAndArticleUserTable extends Migration
             $table->string('content', 100000)->nullable();
             $table->string('image_full');
             $table->string('image_thumb');
-            $table->integer('published')->default(0);
+            $table->tinyInteger('published')->default(0);
             $table->string('views')->nullable();
             $table->timestamps();
         });
